Separate pip search command from operation
By extracting the logic into `pip.operations.search`, the hope is that
folks could do a `search` programmatically more easily.

It also has greater separation of concerns and should allow people to
work in parallel with less chance of merge conflicts.

Continuing work started in #2173 and #2404.
