+++ tox.ini
@@ -21,13 +21,11 @@ commands = check-manifest
 [testenv:pep8]
 basepython = python2.7
 deps = flake8
-       pep8<1.6
 commands = flake8 .
 
 [testenv:py3pep8]
 basepython = python3.3
 deps = flake8
-       pep8<1.6
 commands = flake8 .
 
 [flake8]
